refactor: 🔧 address review — single char-safe truncation helper in its own module
extract_translation_value and extract_config_value each hand-rolled their
own char-based truncation after the byte-slice panic fix, duplicating
hover.rs's existing truncate_for_display — the duplication was the root
cause of the original bug, since the helper was already safe and the
hand-written copies were not.

- Moved truncate_for_display out of hover.rs into its own
  laravel-lsp/src/display_truncate.rs module (declared in lib.rs next to
  the other single-purpose modules), with its three tests moved to
  display_truncate/tests.rs.
- Deleted both inline truncation copies in main.rs; extract_translation_value
  and extract_config_value now call the shared helper, as do the two
  existing hover callers (hover_for_config, hover_for_translation), updated
  to the new module path.
- Consequences of dedup, as approved: the truncation threshold is now
  character-based at the helper's 200-char limit (was a 47/50-char
  threshold) — a 30-char/60-byte Czech string is no longer truncated — and
  the ellipsis at both fixed sites is now the single "…" char, matching
  hover.
- Updated the existing char-boundary regression test in
  byte_offset_panic_hardening.rs for the new limit/ellipsis, added a test
  for the "no longer truncated under 200 chars" consequence, and added the
  equivalent multibyte-boundary regression test for extract_config_value.
